Day5-Production/Assignment/sample_agent/requirements.txt
google-adk
litellm>=1.0.0
opentelemetry-instrumentation-google-genai